Online Exam Management System
Online Exam Management System project that allows students to take online exams through the website and teachers to upload online exams through the website. This system can be accessed by 3 types of users: Student, Teacher, Admin. The student can take the exams uploaded by the teacher and check the results of these exams. The teacher can add or remove exams and check the exam results of his/her students. Admin can add and remove exams, add and remove teachers and students to the system, view the exam results of the students and the exams added by the teachers, view the messages from the mail in the contact section.